[MyCollection] Rename index files
