+++ emission/Pod/Classes/ViewControllers/ARSellTabAppViewController.m
@@ -1,0 +1,10 @@
+#import "ARSellTabAppViewController.h"
+
+@implementation ARSellTabAppViewController
+
+- (instancetype)init
+{
+    return [super initWithEmission:nil moduleName:@"SellTabApp" initialProperties:nil];
+}
+
+@end
