--- emission/Pod/Classes/ViewControllers/ARSellTabLandingViewController.m
@@ -1,10 +1,0 @@
-#import "ARSellTabLandingViewController.h"
-
-@implementation ARSellTabLandingViewController
-
-- (instancetype)init
-{
-    return [super initWithEmission:nil moduleName:@"SellTabLanding" initialProperties:nil];
-}
-
-@end
