+++ src/lib/Scenes/Consignments/index.tsx
@@ -4,7 +4,7 @@ import React from "react"
 import { NativeModules, ViewProperties } from "react-native"
 import NavigatorIOS from "react-native-navigator-ios"
 import Welcome from "./Screens/Welcome"
-import { SellTabLanding } from "./v2"
+import { SellTabApp } from "./v2/SellTabApp"
 
 /** The metadata for a consigned work */
 export interface ConsignmentMetadata {
@@ -67,7 +67,7 @@ interface Props extends ViewProperties, ConsignmentSetup {}
 export default class Consignments extends React.Component<Props, any> {
   render() {
     const featureFlag = NativeModules?.Emission?.options?.AROptionsEnableSales
-    const ConsignmentsEntrypoint = featureFlag ? SellTabLanding : Welcome
+    const ConsignmentsEntrypoint = featureFlag ? SellTabApp : Welcome
     const initialRoute = {
       component: ConsignmentsEntrypoint,
       title: "Welcome",
