+++ src/lib/Scenes/Consignments/v2/Screens/ConsignmentsHome/__tests__/index-tests.tsx
@@ -4,9 +4,9 @@ import React from "react"
 import { create } from "react-test-renderer"
 import { useTracking } from "react-tracking"
 import { createMockEnvironment, MockPayloadGenerator } from "relay-test-utils"
-import { ConsignmentsHomeQueryRenderer } from "../"
 import { ArtistList } from "../Components/ArtistList"
 import { RecentlySold } from "../Components/RecentlySold"
+import { ConsignmentsHomeQueryRenderer } from "../ConsignmentsHome"
 
 jest.unmock("react-relay")
 jest.mock("react-tracking")
