+++ src/lib/Scenes/Consignments/v2/Screens/MyCollectionArtworkDetail/MyCollectionArtworkDetail.tsx
