+++ src/lib/Scenes/Consignments/v2/SellTabApp.tsx
@@ -1,10 +1,12 @@
 import React from "react"
 import { NativeModules } from "react-native"
 import { Boot } from "./Boot"
-import { ConsignmentsHomeQueryRenderer as ConsignmentsHome } from "./Screens/ConsignmentsHome"
-import { MyCollectionHome } from "./Screens/MyCollectionHome"
+import { ConsignmentsHomeQueryRenderer as ConsignmentsHome } from "./Screens/ConsignmentsHome/ConsignmentsHome"
+import { MyCollectionHome } from "./Screens/MyCollectionHome/MyCollectionHome"
 
-export const SellTabLanding: React.FC = () => {
+// TODO: Rename to MyCollectionApp once launched
+
+export const SellTabApp: React.FC = () => {
   const myCollectionEnabled = NativeModules.Emission.options?.AROptionsEnableMyCollection
   const SellTabHome = () => (myCollectionEnabled ? <MyCollectionHome /> : <ConsignmentsHome />)
